Extract accepted connection interface into separate file
