+++ divi/src/rpcserver.h
@@ -24,16 +24,6 @@
 class CBlockIndex;
 class CNetAddr;
 
-class AcceptedConnection
-{
-public:
-    virtual ~AcceptedConnection() {}
-
-    virtual std::iostream& stream() = 0;
-    virtual std::string peer_address_to_string() const = 0;
-    virtual void close() = 0;
-};
-
 /** Start RPC threads */
 void StartRPCThreads();
 /**
